// src/types/general.ts
export interface ReactSelectOptions {
  value: string;
  label: string;
}

export interface AddressType {
  _id: string;
  apartment: string;
  city: string;
  company: string;
  country: string;
  firstName: string;
  lastName: string;
  isDefault: boolean;
  phone: string;
  postalCode: string;
  state: string;
  streetAddress: string;
}